driftstack/requirements.txt

numpy>=1.26
scikit-learn>=1.5
pandas>=2.2
pyyaml>=6.0
tqdm>=4.66
river>=0.21.1
evidently>=0.4.36
mlflow>=2.16.0
loguru>=0.7.2
